add: send long msgs

+++ main/ble_device.h
@@ -2,4 +2,4 @@
 #include <stdlib.h>
 
 void ble_main();
-void ble_send(size_t mapid, size_t report_id, uint8_t *data, uint8_t length);
+void ble_send(size_t mapid, size_t report_id, uint8_t *data, size_t length);
